Moving all the app modules into the "app" folder. using QeAppComputationalResourceWidget for phonopy.
