--- aiidalab_qe_vibroscopy/workflows/__init__.py
@@ -1,25 +1,0 @@
-from aiidalab_qe_vibroscopy.workflows.settings import Setting
-from aiidalab_qe_vibroscopy.workflows.workchain import workchain_and_builder
-from aiidalab_qe_vibroscopy.workflows.result import Result
-from aiidalab_qe.common.panel import OutlinePanel
-
-from aiidalab_widgets_base import ComputationalResourcesWidget
-
-
-class Outline(OutlinePanel):
-    title = "Vibrational properties"
-    # description = "IR and Raman spectra; you may also select phononic and dielectric properties"
-
-
-phonopy_code = ComputationalResourcesWidget(
-    description="phonopy",
-    default_calc_job_plugin="phonopy.phonopy",
-)
-
-property = {
-    "outline": Outline,
-    "code": {"phonopy": phonopy_code},
-    "setting": Setting,
-    "workchain": workchain_and_builder,
-    "result": Result,
-}
